+++ e2e-tests/mdx/gatsby-browser.js
@@ -1,0 +1,6 @@
+const React = require("react")
+const Layout = require("./src/components/layout").default
+
+exports.wrapPageElement = ({ element, props }) => (
+  <Layout {...props}>{element}</Layout>
+)